+++ src/Optimisarr.Data/OptimisarrDbContext.cs
@@ -33,6 +33,7 @@ public sealed class OptimisarrDbContext(DbContextOptions<OptimisarrDbContext> op
             entity.Property(library => library.TargetVideoCodec).HasMaxLength(64);
             entity.Property(library => library.TargetContainer).HasMaxLength(32);
             entity.Property(library => library.ExcludePaths).HasMaxLength(2048);
+            entity.Property(library => library.EncoderPreset).HasMaxLength(32);
         });
 
         modelBuilder.Entity<MediaFile>(entity =>